Better separatation of ASGI
